requirements.txt
torch>=2.5.0
transformers>=4.40.0
datasets>=2.19.0
wandb>=0.17.0
pyyaml>=6.0
pytest>=8.0.0
numpy>=1.26.0
tqdm>=4.66.0
